fix warnings in aigRewriting; only unused functions and veriables used only by assert left

+++ lonestar/aigRewriting/algorithms/ReconvDrivenCut.h
@@ -54,7 +54,7 @@ public:
 
   virtual ~ReconvDrivenCut();
 
-  void run(int cutSizeLimit);
+  void run(size_t cutSizeLimit);
 };
 
 } /* namespace algorithm */
